+++ src/external/matchers/index.ts
@@ -1,1 +1,1 @@
-export { matchesConversationRelevance, Message } from './relevance';
+export { matchesConversationRelevance, Message } from './deepeval';
